Revert "Support local explorer `/cdn-cgi/` routes (#12855)"
This reverts commit c2b76bc35bd038200b9919179f63b66e190e00eb.
